Bypass translated qs machinery if not performing additional filtering
This fixes an issue where an `earliest` or `latest` call is actually a series of queryset calls like: filter, order_by, limit, get (filter). This second `filter` is empty (so has no effect), but occurs after the "slice" (limit) has happened, causing the filter to fail. This fix just skips the translation specific filtering which would be empty anyway, and therefore counts as an empty filter/etc and does not raise an exception for invalid state.