Fix CommandTest for PostgreSQL
Use TransactionTestCase in place of TestCase for CommandTest.

PostgreSQL requires proper handling of transactions to ensure that
connections are set up and torn down in time for this set of tests.
(This had worked for SQLite3 because it does not have the per-connection
overhead of other databases.)

Note that you may encounter "AssertionError: database connection isn't
set to UTC" if you are using a newer version of psycopg2 with an older
version of Django.  Upgrade Django to version 3.1 or newer.  Details:

  https://stackoverflow.com/q/68024060